[INIT] Eureka Server/Client [FEATURE] Moved REST to Eureka Client
