+++ eureka-client/src/main/java/ru/stand/eurekaclient/exception/AuthorIsDeletedException.java
@@ -1,0 +1,7 @@
+package ru.stand.eurekaclient.exception;
+
+public class AuthorIsDeletedException extends Exception{
+    public AuthorIsDeletedException(String message) {
+        super(message);
+    }
+}
