+++ eureka-client/src/main/java/ru/stand/eurekaclient/exception/EmptyPageException.java
@@ -1,0 +1,10 @@
+package ru.stand.eurekaclient.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class EmptyPageException extends Exception{
+    private String message;
+}
