+++ eureka-client/src/main/java/ru/stand/eurekaclient/exception/InvalidBookToAuthorException.java
@@ -1,0 +1,11 @@
+package ru.stand.eurekaclient.exception;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class InvalidBookToAuthorException extends Exception{
+    public String message;
+}
