+++ eureka-client/src/main/java/ru/stand/eurekaclient/exception/NoAuthorFoundException.java
@@ -1,0 +1,14 @@
+package ru.stand.eurekaclient.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.NoSuchElementException;
+
+
+//TODO настроить работу исключения по логике
+@Data
+@AllArgsConstructor
+public class NoAuthorFoundException extends NoSuchElementException {
+    public String message;
+}
