+++ eureka-client/src/main/java/ru/stand/eurekaclient/repository/BookRepository.java
@@ -1,0 +1,13 @@
+package ru.stand.eurekaclient.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+import ru.stand.eurekaclient.entity.Book;
+
+import java.util.List;
+import java.util.UUID;
+
+@Repository
+public interface BookRepository extends JpaRepository<Book, UUID> {
+    List<Book> findBookByName(String name);
+}
