+++ eureka-client/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+server.port = 8081
+
+spring.application.name=eclient
+
+eureka.client.service-url.defaultZone = http://localhost:8080/eureka
+logging.pattern.console=%C{1.} [%-5level] %d{HH:mm:ss} - %msg%n
