+++ eureka-client/src/test/java/ru/stand/eurekaclient/EurekaClientApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.stand.eurekaclient;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaClientApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
