+++ eureka-server/src/main/resources/application.properties
@@ -1,0 +1,17 @@
+server.port = 8080
+
+spring.application.name=eserver
+
+eureka.client.register-with-eureka=false
+eureka.client.fetch-registry=false
+
+
+eureka.client.service-url.defaultZone = http://localhost:8080/eureka
+
+#???? Eureka
+
+logging.level.com.netflix.eureka=OFF
+logging.level.com.netflix.discovery=OFF
+
+logging.pattern.console=%C{1.} [%-5level] %d{HH:mm:ss} - %msg%n
+
