+++ eureka-server/src/test/java/ru/stand/eurekaserver/EurekaServerApplicationTests.java
@@ -1,0 +1,13 @@
+package ru.stand.eurekaserver;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class EurekaServerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
