+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'MVC_practice'
+
+includeBuild 'eureka-server'
+includeBuild 'eureka-client'
