--- src/main/java/ru/k1morng/mvc_practice/dto/BookDto.java
@@ -1,18 +1,0 @@
-package ru.k1morng.mvc_practice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-import java.util.UUID;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@ToString
-public class BookDto {
-    private UUID id;
-    private String name;
-    private int pages;
-}
