--- src/main/java/ru/k1morng/mvc_practice/entity/Book.java
@@ -1,46 +1,0 @@
-package ru.k1morng.mvc_practice.entity;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.hibernate.Hibernate;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.util.Objects;
-import java.util.UUID;
-
-@Entity
-@Table(name = "book")
-@Getter
-@Setter
-@NoArgsConstructor
-public class Book {
-
-    @Id
-    private UUID id;
-    @Column(name = "name")
-    private String name;
-    @Column(name = "pages")
-    private int pages;
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) {
-            return false;
-        }
-        Book book = (Book) o;
-        return getId() != null && Objects.equals(getId(), book.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
