--- src/main/java/ru/k1morng/mvc_practice/exception/AuthorIsDeletedException.java
@@ -1,7 +1,0 @@
-package ru.k1morng.mvc_practice.exception;
-
-public class AuthorIsDeletedException extends Exception{
-    public AuthorIsDeletedException(String message) {
-        super(message);
-    }
-}
