--- src/main/java/ru/k1morng/mvc_practice/exception/EmptyPageException.java
@@ -1,10 +1,0 @@
-package ru.k1morng.mvc_practice.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@AllArgsConstructor
-@Data
-public class EmptyPageException extends Exception{
-    private String message;
-}
