--- src/main/java/ru/k1morng/mvc_practice/mapper/BookMapper.java
@@ -1,26 +1,0 @@
-package ru.k1morng.mvc_practice.mapper;
-
-import org.mapstruct.InheritInverseConfiguration;
-import org.mapstruct.InjectionStrategy;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-import org.mapstruct.factory.Mappers;
-import ru.k1morng.mvc_practice.dto.BookDto;
-import ru.k1morng.mvc_practice.entity.Book;
-
-import java.util.UUID;
-
-@Mapper(componentModel = "spring",
-        injectionStrategy = InjectionStrategy.FIELD,
-        imports = {UUID.class})
-public abstract class BookMapper {
-
-    public static final BookMapper INSTANCE = Mappers.getMapper(BookMapper.class);
-
-    @Mapping(target = "id", expression = "java(UUID.randomUUID())")
-    public abstract Book fromDto(BookDto bookDto);
-
-    @InheritInverseConfiguration
-    public abstract BookDto toBookDto(Book book);
-
-}
