--- src/main/java/ru/k1morng/mvc_practice/repository/BookRepository.java
@@ -1,13 +1,0 @@
-package ru.k1morng.mvc_practice.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-import ru.k1morng.mvc_practice.entity.Book;
-
-import java.util.List;
-import java.util.UUID;
-
-@Repository
-public interface BookRepository extends JpaRepository<Book, UUID> {
-    List<Book> findBookByName(String name);
-}
