--- src/main/resources/application.yaml
@@ -1,45 +1,0 @@
-server:
-  port: 8080
-  servlet:
-    encoding:
-      force-response: true
-      charset: UTF-8
-
-spring:
-  main:
-    banner-mode: off
-    allow-bean-definition-overriding: true
-    allow-circular-references: true
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    username: postgres
-    password: postgres
-    url: jdbc:postgresql://localhost:5432/postgres
-  jpa:
-    open-in-view: false
-    hibernate:
-      ddl-auto: validate
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.PostgreSQL95Dialect
-        hbm2ddl:
-          auto: validate
-  mandatory-file-encoding: UTF-8
-  liquibase:
-    enabled: true
-    change-log: 'classpath:liquibase/changelog-master.yml'
-logging:
-  level:
-    root: warn
-    ru: debug
-management:
-  server:
-    port: 8081
-  endpoints:
-    web:
-      base-path: /actuator
-      exposure:
-        include: "*"
-  endpoint:
-    health:
-      show-details: always
